llm_apps/requirements.txt
langchain
langchain-community
python-dotenv
openai
google-search-results  # This is required for SerpAPI